XMLHttpRequest/withcredentials-anonymous.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: setting withCredentials on anonymous and sync request should throw</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#the-withcredentials-attribute" data-tested-assertations="following::ol/li[3] following::ol/li[4]" />
    <link rel="help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#the-open()-method" data-tested-assertations="following::ol/li[10]" />
  </head>
  <body>
    <div id="log"></div>
    <script>
      var client = new XMLHttpRequest({anonymous:true})
      test(function() {
        client.open("GET", "resources/delay.py?ms=1000")
        assert_throws("InvalidStateError", function() { client.withCredentials=true })
      }, "throw when set on anonymous request")

      var client = new XMLHttpRequest()
      test(function() {
        client.open("GET", "resources/delay.py?ms=1000", false)
        assert_throws("InvalidStateError", function() { client.withCredentials=true })
      }, "throw when set on sync request")

      var client = new XMLHttpRequest()
      test(function() {
        client.withCredentials=true
        assert_throws("InvalidStateError", function() { client.open("GET", "resources/delay.py?ms=1000", false) })
      }, "throw when opening sync request after setting withCredentials to true")

    </script>
  </body>
</html>
